combined the DatabaseSetup into one filed
